Added the ability to add a board to "favorites". Added the ability to change the board name.

+++ NoteCloud_Config/static/boards/js/trash.js
@@ -36,10 +36,61 @@ document.querySelector('.main_block').addEventListener('click', (e) => {
     if (deleteBtn) {
         const csrftoken = getCookie('csrftoken');
         const boardElement = deleteBtn.closest('.board_item');
-        const urlHash = boardElement.querySelector('a').getAttribute('href').split('/')[2];
+        const urlHash = deleteBtn.getAttribute('data-url_hash');
 
-        fetch(`/workspace/trash/${urlHash}/delete/`, {
-            method: 'DELETE',
+        if (confirm('Вы уверены, что хотите удалить эту доску?')) {
+            fetch(`/workspace/trash/${urlHash}/delete/`, {
+                method: 'DELETE',
+                headers: {
+                    'X-CSRFToken': csrftoken
+                }
+            })
+            .then(response => response.json())
+            .then(data => {
+                if (data.status === 'success') {
+                    boardElement.remove();
+                } else {
+                    alert('Ошибка при удалении доски');
+                }
+            });
+        }
+    }
+});
+
+document.getElementById("delete-all-btn").addEventListener("click", function(){
+    if (!confirm('Вы уверены, что хотите удалить все ваши доски?')) {
+      return;
+    }
+    const csrftoken = getCookie('csrftoken');
+    const currentUsername = window.currentUsername;
+    fetch(`/workspace/trash/delete_all/`, {
+      method: "DELETE",
+      headers: {
+        "X-CSRFToken": csrftoken
+      }
+    })
+    .then(response => response.json())
+    .then(data => {
+      if(data.status === "success"){
+        // Удаляем со страницы все доски пользователя
+        document.querySelectorAll('.board_item').forEach(function(item) {
+            item.remove();
+        });
+      }
+    })
+    .catch(error => console.error('Ошибка:', error));
+});
+
+
+document.querySelector('.main_block').addEventListener('click', (e) => {
+    const restoreBtn = e.target.closest('.restore-btn');
+    if (restoreBtn) {
+        const csrftoken = getCookie('csrftoken');
+        const boardElement = restoreBtn.closest('.board_item');
+        const urlHash = restoreBtn.getAttribute('data-url_hash');
+
+        fetch(`/workspace/trash/${urlHash}/restore/`, {
+            method: 'POST',
             headers: {
                 'X-CSRFToken': csrftoken
             }
@@ -53,4 +104,25 @@ document.querySelector('.main_block').addEventListener('click', (e) => {
             }
         });
     }
+});
+
+document.getElementById("restore-all-btn").addEventListener("click", function(){
+    const csrftoken = getCookie('csrftoken');
+    const currentUsername = window.currentUsername;
+    fetch(`/workspace/trash/restore_all/`, {
+      method: "POST",
+      headers: {
+        "X-CSRFToken": csrftoken
+      }
+    })
+    .then(response => response.json())
+    .then(data => {
+      if(data.status === "success"){
+        // Удаляем со страницы все доски пользователя
+        document.querySelectorAll('.board_item').forEach(function(item) {
+            item.remove();
+        });
+      }
+    })
+    .catch(error => console.error('Ошибка:', error));
 });
